requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
docx==0.2.4
greenlet==3.2.2
lxml==5.4.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
playwright==1.52.0
pyee==13.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
